pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qiskit-state-evolution-recorder"
version = "0.0.0"
description = "Simple module allowing to record animations to trace changes in qubit states for arbitrary quantum circuits."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name = "Dawid Ciepiela", email = "71898979+sarumaj@users.noreply.github.com" }
]
requires-python = ">=3.6, <3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "numpy==2.2.6",
    "qiskit>=1.4.3,<=2.1.0",
    "pillow==11.2.1",
    "pylatexenc==2.10",
    "tqdm==4.67.1",
]

[project.optional-dependencies]
opencv = [
    "opencv-python==4.11.0.86"
]
test = [
    "psutil==7.0.0",
    "pytest==8.4.0",
    "pytest-benchmark==5.1.0"
]

[project.scripts]
qiskitrecorder = "qiskit_state_evolution_recorder.__main__:main"

[project.urls]
"Homepage" = "https://github.com/sarumaj/qiskit-state-evolution-recorder"

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"qiskit_state_evolution_recorder" = [
    "demo.qasm"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = ["-v", "--benchmark-disable"]
filterwarnings = [
    "ignore:This process .* is multi-threaded, use of fork\\(\\) may lead to deadlocks in the child:DeprecationWarning:multiprocessing.popen_fork",
]